Add create and update function

+++ DAL/Product.cs
@@ -23,7 +23,7 @@ namespace DAL
         public int Id { get; set; }
         public string Name { get; set; }
         public string Number { get; set; }
-        public double Cost { get; set; }
+        public int Cost { get; set; }
         public int Category_FK { get; set; }
     
         public virtual Category Category { get; set; }
